managed-connectivity/community-contributed-connectors/databricks-connector/requirements.txt
google-cloud-dataplex
google-cloud-storage
google-cloud-secret-manager
google-cloud-logging
cryptography
databricks-sql-connector